Update README.md
# PMES Theory Website

This repository contains the source code for the official website of **PMES Theory** (Photon-Membrane & Equilibrium Singularity).

## About PMES Theory
PMES is a novel theoretical framework that models reality as excitations of a fundamental field Ψ in a 12-dimensional manifold M = A × B. The theory provides geometric explanations for:
- Emergence of spacetime and gravity
- Quantum phenomena and particle masses
- Structure of photons as 11D membrane-solitons
- Resolution of black hole singularities

## Live Website
Visit the official site: [https://pmes-theory.org](https://pmes-theory.org)

## Publications
All research papers are available on Zenodo. Search for "PMES" or "[Your Name]".

## License
This website content is licensed under the MIT License.